Fix
Исправил названия, добавил тип, отступы, сделал cleanup и для namespace, добавил action
