+++ WarehouseEnterprise/TESTS/WarehouseEnterprise.Tests.csproj
@@ -17,7 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\WarehouseEnterpriseApp\WarehouseEnterpriceApp.csproj" />
+    <ProjectReference Include="..\WarehouseEnterpriseApp\WarehouseEnterpriseApp.csproj" />
   </ItemGroup>
 
   <ItemGroup>
